Add firstname last name. Refactor into smaller files.
