+++ quiz-us-components/src/components/MasteryMap/README.md
@@ -2,7 +2,6 @@
 
 ## Description
 
-Rich text editor built using Slate.js
 
 ## Usage
