+++ quiz-us-components/src/components/MasteryMap/index.js
@@ -1,166 +1,59 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+
 import { lighten, makeStyles } from '@material-ui/core/styles';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
 import TablePagination from '@material-ui/core/TablePagination';
 import TableRow from '@material-ui/core/TableRow';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
 import Paper from '@material-ui/core/Paper';
-import IconButton from '@material-ui/core/IconButton';
-import Tooltip from '@material-ui/core/Tooltip';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Switch from '@material-ui/core/Switch';
-import FilterListIcon from '@material-ui/icons/FilterList';
-import EnhancedTableHead from './components/TableHead';
 
-function createData(name, standard1, standard2, standard3, standard4) {
-    return { name, standard1, standard2, standard3, standard4 };
+import EnhancedTableHead from './components/TableHead';
+import { stableSort, getSorting } from "../MasteryMap/utils";
+import EnhancedTableToolbar from './components/EnhancedTableToolbar';
+
+//key name needs to be the same as the readrows id
+function createData(id, lastname, firstname, standard1, standard2, standard3, standard4) {
+  return { id, lastname, firstname, standard1, standard2, standard3, standard4 };
 }
 
 const rows = [
-    createData('StudentA', 100, 100, 110, 100),
-    createData('StudentB', 80, 80, 80, 80),
-    createData('StudentC', 60, 60, 60, 60),
-    createData('StudentD', 40, 40, 40, 40),
-    createData('StudentE', 20, 20, 20, 20),
-    createData('StudentF', 0, 0, 0, 0),
+  createData(1, 'A', 'StudentA', null, 100, 110, 100),
+  createData(2, 'B', 'StudentB', 80, 70, 20, 80),
+  createData(3, 'C', 'StudentC', 60, 20, 60, 90),
+  createData(4, 'D', 'StudentD', 100, 40, 50, 40),
+  createData(5, 'E', 'StudentE', 20, 20, 70, 20),
+  createData(6, 'F', 'StudentF', 0, 0, 0, 0),
 ];
 
-function desc(a, b, orderBy) {
-    if (b[orderBy] < a[orderBy]) {
-        return -1;
-    }
-    if (b[orderBy] > a[orderBy]) {
-        return 1;
-    }
-    return 0;
-}
-
-function stableSort(array, cmp) {
-    const stabilizedThis = array.map((el, index) => [el, index]);
-    stabilizedThis.sort((a, b) => {
-        const order = cmp(a[0], b[0]);
-        if (order !== 0) return order;
-        return a[1] - b[1];
-    });
-    return stabilizedThis.map(el => el[0]);
-}
-
-function getSorting(order, orderBy) {
-    return order === 'desc' ? (a, b) => desc(a, b, orderBy) : (a, b) => -desc(a, b, orderBy);
-}
-
+//key name needs to be the same as the readrows id
 const headRows = [
-    { id: 'Name', numeric: false, disablePadding: false, label: 'Name' },
-    { id: 'standard1', numeric: true, disablePadding: false, label: 'Standard1' },
-    { id: 'standard2', numeric: true, disablePadding: false, label: 'Standard2' },
-    { id: 'standard3', numeric: true, disablePadding: false, label: 'Standard3' },
-    { id: 'standard4', numeric: true, disablePadding: false, label: 'Standard4' },
+  { id: 'lastname', numeric: false, disablePadding: false, label: 'Last' },
+  { id: 'firstname', numeric: false, disablePadding: false, label: 'First' },
+  { id: 'standard1', numeric: true, disablePadding: false, label: 'Standard1' },
+  { id: 'standard2', numeric: true, disablePadding: false, label: 'Standard2' },
+  { id: 'standard3', numeric: true, disablePadding: false, label: 'Standard3' },
+  { id: 'standard4', numeric: true, disablePadding: false, label: 'Standard4' },
 ];
-
-// function EnhancedTableHead(props) {
-//     const { order, orderBy, onRequestSort } = props;
-//     const createSortHandler = property => event => {
-//         onRequestSort(event, property);
-//     };
-
-//     return (
-//         <TableHead>
-//             <TableRow>
-//                 {headRows.map(row => (
-//                     <TableCell
-//                         key={row.id}
-//                         align={row.numeric ? 'right' : 'left'}
-//                         padding={row.disablePadding ? 'none' : 'default'}
-//                         sortDirection={orderBy === row.id ? order : false}
-//                     >
-//                         <TableSortLabel
-//                             active={orderBy === row.id}
-//                             direction={order}
-//                             onClick={createSortHandler(row.id)}
-//                         >
-//                             {row.label}
-//                         </TableSortLabel>
-//                     </TableCell>
-//                 ))}
-//             </TableRow>
-//         </TableHead>
-//     );
-// }
-
-// EnhancedTableHead.propTypes = {
-//     onRequestSort: PropTypes.func.isRequired,
-//     order: PropTypes.string.isRequired,
-//     orderBy: PropTypes.string.isRequired,
-//     rowCount: PropTypes.number.isRequired,
-// };
-
-const useToolbarStyles = makeStyles(theme => ({
-    root: {
-        paddingLeft: theme.spacing(2),
-        paddingRight: theme.spacing(1),
-    },
-    highlight:
-        theme.palette.type === 'light'
-            ? {
-                color: theme.palette.secondary.main,
-                backgroundColor: lighten(theme.palette.secondary.light, 0.85),
-            }
-            : {
-                color: theme.palette.text.primary,
-                backgroundColor: theme.palette.secondary.dark,
-            },
-    spacer: {
-        flex: '1 1 100%',
-    },
-    actions: {
-        color: theme.palette.text.secondary,
-    },
-    title: {
-        flex: '0 0 auto',
-    },
-}));
-
-const EnhancedTableToolbar = props => {
-    const classes = useToolbarStyles();
-
-    return (
-        <Toolbar>
-            <div className={classes.title}>
-                <Typography variant="h6" id="tableTitle">
-                    MasteryMap
-        </Typography>
-            </div>
-            <div className={classes.spacer} />
-            <div className={classes.actions}>
-                <Tooltip title="Filter list">
-                    <IconButton aria-label="Filter list">
-                        <FilterListIcon />
-                    </IconButton>
-                </Tooltip>
-            </div>
-        </Toolbar>
-    );
-};
 
 const useStyles = makeStyles(theme => ({
-    root: {
-        width: '100%',
-        marginTop: theme.spacing(3),
-    },
-    paper: {
-        width: '100%',
-        marginBottom: theme.spacing(2),
-    },
-    table: {
-        minWidth: 750,
-    },
-    tableWrapper: {
-        overflowX: 'auto',
-    },
+  root: {
+    width: '100%',
+    marginTop: theme.spacing(3),
+  },
+  paper: {
+    width: '100%',
+    marginBottom: theme.spacing(2),
+  },
+  table: {
+    minWidth: 750,
+  },
+  tableWrapper: {
+    overflowX: 'auto',
+  },
 }));
 
 export default function EnhancedTable() {
@@ -172,91 +65,95 @@ export default function EnhancedTable() {
     const [rowsPerPage, setRowsPerPage] = React.useState(10);
 
     function handleRequestSort(event, property) {
-        const isDesc = orderBy === property && order === 'desc';
-        setOrder(isDesc ? 'asc' : 'desc');
-        setOrderBy(property);
+      console.log(event, property);
+      const isDesc = orderBy === property && order === 'desc';
+      setOrder(isDesc ? 'asc' : 'desc');
+      setOrderBy(property);
     }
     function handleChangePage(event, newPage) {
-        setPage(newPage);
+      setPage(newPage);
     }
 
     function handleChangeRowsPerPage(event) {
-        setRowsPerPage(+event.target.value);
+      setRowsPerPage(+event.target.value);
     }
 
     function handleChangeDense(event) {
-        setDense(event.target.checked);
+      setDense(event.target.checked);
     }
 
-    const emptyRows = rowsPerPage - Math.min(rowsPerPage, rows.length - page * rowsPerPage);
+    // const emptyRows = rowsPerPage - Math.min(rowsPerPage, rows.length - page * rowsPerPage);
 
     return (
-        <div className={classes.root}>
-            <Paper className={classes.paper}>
-                <EnhancedTableToolbar />
-                <div className={classes.tableWrapper}>
-                    <Table
-                        className={classes.table}
-                        aria-labelledby="tableTitle"
-                        size={dense ? 'small' : 'medium'}
-                    >
-                        <EnhancedTableHead
-                            headRows={headRows}
-                            order={order}
-                            orderBy={orderBy}
-                            onRequestSort={handleRequestSort}
-                            rowCount={rows.length}
-                        />
-                        <TableBody>
-                            {stableSort(rows, getSorting(order, orderBy))
-                                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                                .map((row, index) => {
-                                    const labelId = `enhanced-table-checkbox-${index}`;
-                                    return (
-                                        <TableRow
-                                            hover
-                                            tabIndex={-1}
-                                            key={row.name}
-                                        >
-                                            <TableCell component="th" id={labelId} scope="row">
-                                                {row.name}
-                                            </TableCell>
-                                            {/* <TableCell>{row.name}</TableCell> */}
-                                            <TableCell align="right">{row.standard1}</TableCell>
-                                            <TableCell align="right">{row.standard2}</TableCell>
-                                            <TableCell align="right">{row.standard3}</TableCell>
-                                            <TableCell align="right">{row.standard4}</TableCell>
-                                        </TableRow>
-                                    );
-                                })}
-                            {emptyRows > 0 && (
-                                <TableRow style={{ height: 49 * emptyRows }}>
-                                    <TableCell colSpan={6} />
-                                </TableRow>
-                            )}
-                        </TableBody>
-                    </Table>
-                </div>
-                <TablePagination
-                    rowsPerPageOptions={[5, 10, 25]}
-                    component="div"
-                    count={rows.length}
-                    rowsPerPage={rowsPerPage}
-                    page={page}
-                    backIconButtonProps={{
-                        'aria-label': 'Previous Page',
-                    }}
-                    nextIconButtonProps={{
-                        'aria-label': 'Next Page',
-                    }}
-                    onChangePage={handleChangePage}
-                    onChangeRowsPerPage={handleChangeRowsPerPage}
-                />
-            </Paper>
-            <FormControlLabel
-                control={<Switch checked={dense} onChange={handleChangeDense} />}
-                label="Dense padding"
-            />
-        </div>
+      <div className={classes.root}>
+        <Paper className={classes.paper}>
+          <EnhancedTableToolbar />
+          <div className={classes.tableWrapper}>
+            <Table
+              className={classes.table}
+              aria-labelledby="tableTitle"
+              size={dense ? 'small' : 'medium'}
+            >
+              <EnhancedTableHead
+                  headRows={headRows}
+                  order={order}
+                  orderBy={orderBy}
+                  onRequestSort={handleRequestSort}
+                  rowCount={rows.length}
+              />
+                <TableBody>
+                  {stableSort(rows, getSorting(order, orderBy))
+                    .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+                    .map((row, index) => {
+                      const labelId = `enhanced-table-checkbox-${index}`;
+                      return (
+                          <TableRow
+                              hover
+                              tabIndex={-1}
+                              key={row.id}
+                          >
+                            {Object.keys(row).map((el,i) => {
+                              if (i === 0) return;
+                              if (i === 1 || i === 2) {
+                                return <TableCell component="th" id={labelId} scope="row">
+                                  {row[el]}
+                                </TableCell>
+                              } else {
+                                return <TableCell align="right">{row[el]}</TableCell>
+                              }
+                            })}
+                          </TableRow>
+                      );
+                  })}
+                  {/* if you want the mastery map to be the same size when there are less rows */}
+                  {/* {emptyRows > 0 && (
+                    <TableRow style={{ height: 49 * emptyRows }}>
+                      <TableCell colSpan={6} />
+                    </TableRow>
+                  )} */}
+                </TableBody>
+              </Table>
+          </div>
+          <TablePagination
+              rowsPerPageOptions={[5, 10, 25]}
+              component="div"
+              count={rows.length}
+              rowsPerPage={rowsPerPage}
+              page={page}
+              backIconButtonProps={{
+                  'aria-label': 'Previous Page',
+              }}
+              nextIconButtonProps={{
+                  'aria-label': 'Next Page',
+              }}
+              onChangePage={handleChangePage}
+              onChangeRowsPerPage={handleChangeRowsPerPage}
+          />
+        </Paper>
+        <FormControlLabel
+            control={<Switch checked={dense} onChange={handleChangeDense} />}
+            label="Dense padding"
+        />
+      </div>
     );
 }
